Moved business logic to logic module
