+++ UI/UILogic.py
@@ -1,14 +1,10 @@
 from kivy.uix.dropdown import DropDown
-from kivy.uix.label import Label
 from kivy.uix.button import Button
-from kivy.uix.image import Image
-from kivy.uix.behaviors import ButtonBehavior
 
 
 class UILogic:
-    def __init__(self, currencyDict, ApiResponse):
+    def __init__(self, currencyDict):
         self.currencyDict = currencyDict
-        self.ApiResponse = ApiResponse
 
     def DropDownCurrencyButton(self, defaultCurrency):
         dropdown = DropDown()
@@ -22,22 +18,3 @@ class UILogic:
         mainButton.bind(on_release=dropdown.open)
         dropdown.bind(on_select=lambda instance, x: setattr(mainButton, 'text', x))
         return mainButton
-
-    def GetConvertedValue(self, currencyFrom, currencyTo, amount):
-        # TODO - should work with every currency (that function works only if currencyFrom is EUR)
-        return amount * self.ApiResponse['rates'][currencyTo]
-
-    def GetUpdateDate(self):
-        # TODO - get date and time
-        return f"Last update: {self.ApiResponse['date']}"
-
-    def Get1EuroInDollars(self):
-        value = self.GetConvertedValue('EUR', 'USD', 1)
-        return f"1 EUR = {value} USD"
-
-    def Get1DollarInEuros(self):
-        value = self.GetConvertedValue('EUR', 'USD', 1)
-        return f"1 USD = {value} EUR"
-
-    def GetConvertedValueString(self, currencyFrom, currencyTo, amount):
-        return f"{amount} {currencyFrom} = {self.GetConvertedValue(currencyFrom, currencyTo, amount)} {currencyTo}"
